+++ python/src/fem.cpp
@@ -193,7 +193,7 @@ void fem(py::module& m)
            py::arg("check_midpoint"))
       .def(py::init<std::shared_ptr<const dolfin::function::FunctionSpace>,
                     std::shared_ptr<const dolfin::function::Function>,
-                    const std::vector<std::size_t>&,
+                    const std::vector<std::int32_t>&,
                     dolfin::fem::DirichletBC::Method>(),
            py::arg("V"), py::arg("g"), py::arg("facets"), py::arg("method"))
       .def("function_space", &dolfin::fem::DirichletBC::function_space)
